+++ .golangci.yml
@@ -43,85 +43,8 @@ linters-settings:
 linters:
   disable-all: true
   enable:
-    - asciicheck
-    - bidichk
-    - bodyclose
-    - contextcheck
-    #    - cyclop
-    - deadcode
-    - depguard
-    - dogsled
-    #    - dupl
-    - durationcheck
-    - errcheck
-    #    - errname
-    #    - errorlint
-    #    - exhaustive
-    #    - exhaustivestruct
-    - exportloopref
-    #    - forbidigo
-    #    - forcetypeassert
-    - funlen
-    #    - gci
-    #    - gochecknoglobals
-    #    - gochecknoinits
-    #    - gocognit
-    - goconst
-    #    - gocritic
-    #    - gocyclo
-    #    - godot
-    #    - godox
-    #    - goerr113
     - gofmt
-    #    - gofumpt
-    - goheader
     - goimports
-    #    - gomnd
-    - gomoddirectives
-    - gomodguard
-    - goprintffuncname
-    #    - gosec
-    - gosimple
-    - govet
-    #    - ifshort
-    - importas
-    - ineffassign
-    #    - ireturn
-    #    - lll
-    - makezero
-    #    - misspell
-    - nakedret
-    - nestif
-    - nilerr
-    #    - nilnil
-    #    - nlreturn
-    - noctx
-    #    - nolintlint
-    #    - paralleltest
-    - prealloc
-    - predeclared
-    - promlinter
-    - revive
-    - rowserrcheck
-    - sqlclosecheck
-    - staticcheck
-    - structcheck
-    - stylecheck
-    #    - tagliatelle
-    - tenv
-    #    - testpackage
-    #    - thelper
-    - tparallel
-    - typecheck
-    - unconvert
-    - unparam
-    - unused
-    - varcheck
-    #    - varnamelen
-    - wastedassign
-    #    - whitespace
-    #    - wrapcheck
-    #    - wsl
 
 issues:
   exclude-rules:
